requirements.txt
django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-environ>=0.11.0
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.7
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
python-decouple>=3.8
whitenoise>=6.6.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.2
coverage>=7.3.0

# Development
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
